Extract Javascript into separate file
