+++ webpack.mix.js
@@ -93,6 +93,10 @@ mix.js("resources/js/app.js", "public/js")
     .js(
         "resources/js/clients/payments/eway-credit-card.js",
         "public/js/clients/payments/eway-credit-card.js"
+    )
+    .js(
+        "resources/js/clients/payment_methods/braintree-ach.js",
+        "public/js/clients/payment_methods/braintree-ach.js"
     );
 
 mix.copyDirectory('node_modules/card-js/card-js.min.css', 'public/css/card-js.min.css');
